backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4

openai==1.58.1
anthropic==0.42.0

chromadb==0.5.23

requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
lxml==5.3.0

apscheduler==3.10.4

python-multipart==0.0.20
